Entires stuff Entire numbers type changed from uint8 to int; Added "bool GetNearEntireCoords(int& entire, uint16& hexX, uint16& hexY) const"; Entire number -1 checked as any entire.
git-svn-id: https://cvet.by:8443/svn/FOnlineSource@3 c2c6a35e-ad2b-4e1f-bb64-97b989be360b

Former-commit-id: 68f6190b86a965bcef08496ea213dd24590fa645

+++ Source/ProtoMap.h
@@ -238,7 +238,7 @@ public:
 	MapEntire* GetEntire(DWORD num, DWORD skip);
 	MapEntire* GetEntireRandom(DWORD num);
 	MapEntire* GetEntireNear(DWORD num, WORD hx, WORD hy);
-	MapEntire* GetEntireNear(DWORD num, BYTE num_ext, WORD hx, WORD hy);
+	MapEntire* GetEntireNear(DWORD num, DWORD num_ext, WORD hx, WORD hy);
 	void GetEntires(DWORD num, EntiresVec& entires);
 
 private:
